+++ controllers/iterator.js
@@ -16,18 +16,19 @@
 // ===========================================================================
 
     module.exports = function(route, signal, callback, state){
-    	if(route.length){
+    	if(route && route.length){
     		var iterator = new ArrayIterator(route);
     		
     		function nextRoute(){
-    			
-    			var current = iterator.next()
-    			
-    			if(!current.done){
-    				current.value.handler(signal.request, signal.response, signal.app, nextRoute, signal)
-    			
-    			} else if (callback) {
-    				callback(signal)
+    			if(!signal.stopped){
+        			var current = iterator.next()
+        			
+        			if(!current.done){
+        				current.value.handler(signal.request, signal.response, signal.app, nextRoute, signal)
+        			
+        			} else if (callback) {
+        				callback(signal)
+        			}
     			}
     		}