+++ controllers/protocols/http/signal.js
@@ -39,7 +39,7 @@
     		},
     		headers: request.headers,	
     		send:  function(message) { response.write(message); },               // send data chunk to client
-    		redirect: function(input, statusCode){
+    		redirect: function(input, statusCode, isLast){
     			if(input.substring(0, 4) === 'back') { 
     				var path = request.headers.referer || '/';
     			} else if(input.substring(0, 4) === 'home') { 
@@ -64,7 +64,7 @@
     			signal.header('Location', path);
     			response.end()
     			signal.responded = true
-    			signal.nextRoute()
+    			if(!isLast) signal.nextRoute()
     		},
     		env: process.env.NODE_ENV,
     		setFinalHeaders: function(data){
@@ -76,14 +76,14 @@
     		    	}
     		    }
     		},
-    		end : function(input){
-    			if(!signal.responded){
+    		end : function(input, isLast){
+    			if(!signal.responded && !signal.stopped){
     			    signal.responded = true
     			    if(typeof input == 'object' || signal.header('x-requested-with') == 'XMLHttpRequest' || ( signal.header('authorization') && (signal.header('authorization').indexOf('Bearer') != -1 || signal.header('authorization').indexOf('Token') != -1 ))) { 
     			        let data = signal.jsonString(input); // json
     			        signal.setFinalHeaders(data); 
     			        response.end(data)      
-    			        signal.nextRoute() // call next route                                                       
+    			        if(!isLast) signal.nextRoute() // call next route                                                       
     			    
     			    } else if (app.html) {
     			        signal.html(input) // html                  
@@ -91,9 +91,10 @@
     			    } else {
                         signal.setFinalHeaders(input); 
                         response.end(input)  // default
-                        signal.nextRoute() // call next route
+                        if(!isLast) signal.nextRoute() // call next route
     			    }
     			}
+    			if(isLast) signal.stopped = true;
     		},
     		status : function(code, message){
     			signal.statusMessage = message || signal.statusMessage || status_codes[code];	
@@ -104,20 +105,20 @@
     			signal.passed = signal.data.passed = false
     			signal.errors[field] = error
     		},
-    		success: function(input){                                       // respond with JSON success
+    		success: function(input, isLast){ // respond with JSON success
     			if(!signal.statusCode) signal.status(200)
     			signal.header('content-type', 'application/json')
     			var data = signal.data
     			if(isset(input)) data = Object.merge(signal.data, input)
     			data.passed = true
-    			signal.end(data)
+    			signal.end(data, isLast)
     		},
-    		failure: function(input){                                            // respond with JSON errors
+    		failure: function(input, isLast){ // respond with JSON errors
     			if(!signal.statusCode) signal.status(200)
     			signal.header('content-type', 'application/json')
     			if(signal.data.errors) signal.errors = Object.merge(signal.error, signal.data.errors)
     			if(isset(input)) data = Object.merge(signal.errors, input)
-    			signal.end({ passed: false, errors: signal.errors })
+    			signal.end({ passed: false, errors: signal.errors }, isLast)
     		},
     		jsonString: function(input){
     		    if(isset(input)) signal.data = Object.merge(signal.data, input)
@@ -125,17 +126,17 @@
     		    signal.header('content-type', 'application/json')
     		    return JSON.stringify(signal.data);
     		},
-    		json : function(input){                                    // respond with JSON data
-    			signal.end(signal.jsonString(input))
+    		json : function(input, isLast){ // respond with JSON data
+    			signal.end(signal.jsonString(input), isLast)
     		},
-    		html: function(input){
+    		html: function(input, isLast){
     		    signal.header('Content-Type', 'text/html; charset=UTF-8')
     		    if(!signal.statusCode) signal.status(200)
     		    if(signal.htmlModule) { 
     		        signal.htmlModule(input) 
     		    } else { 
     		        response.end(input) 
-    		        signal.nextRoute() // call next route
+    		        if(!isLast) signal.nextRoute() // call next route
     		    }
     		}
     	}